[ALPS02695402] AEE: bug fix for 64bit projects with no backtrace in HWT
[Detail]
return type of unsigned int of get_linear_memory_size() function
will cut off the high byte value thus it will return 0 which cause
the check of aee_dump_stack_top_binary() function
to return failure status.

[Solution]
change unsigned int to unsigned long

MTK-Commit-Id: e70252cf1065370cac8b5d06e5dc03657f5ca3a6

CR-Id: ALPS02695402
Feature: Android Exception Engine(AEE)
(cherry picked from commit 839ae6d2703df8cc5fc18b54395dd057657df6cc)

+++ drivers/misc/mediatek/aee/common/reboot-reason.c
@@ -148,7 +148,7 @@ void ksysfs_bootinfo_exit(void)
 
 /* end sysfs bootinfo */
 
-static inline unsigned int get_linear_memory_size(void)
+static inline unsigned long get_linear_memory_size(void)
 {
 	return (unsigned long)high_memory - PAGE_OFFSET;
 }
